correct user update and not coorect film adding and updating

+++ src/main/java/ru/yandex/practicum/filmorate/controller/GenreController.java
@@ -32,7 +32,7 @@ public class GenreController {
     }
 
     @GetMapping("/{id}")
-    public Genre getGenreById(@Valid @PathVariable("id") @Positive Long id) {
+    public Genre getGenreById(@Valid @PathVariable("id") @Positive Integer id) {
         return genreService.getGenreById(id);
     }
 }
