+++ src/main/java/ru/yandex/practicum/filmorate/dal/mappers/UserRowMapper.java
@@ -12,7 +12,7 @@ public class UserRowMapper implements RowMapper<User> {
     @Override
     public User mapRow(ResultSet rs, int rowNum) throws SQLException {
         return User.builder()
-                .id(rs.getLong("id"))
+                .id(rs.getInt("id"))
                 .name(rs.getString("name"))
                 .email(rs.getString("email"))
                 .login(rs.getString("login"))
